Refactor subtests to separate files
