+++ setup.cfg
@@ -41,10 +41,8 @@ test =
     pytest-cov
 
 [tool:pytest]
-testpaths = "tests"
-markers =
-    image: run test during image comparison only mode.
-    hash: run test during hash comparison only mode.
+testpaths = tests
+norecursedirs = tests/subtests/subtest
 filterwarnings =
     error
     ignore:distutils Version classes are deprecated
