+++ tests/subtests/subtest/baseline/test_hdiff_idiff.png
