+++ tests/subtests/subtest/baseline/test_hdiff_idiff_tolerance.png
