+++ tests/subtests/subtest/baseline/test_hdiff_idiffshape.png
