+++ tests/subtests/subtest/baseline/test_hdiff_imatch.png
