+++ tests/subtests/subtest/baseline/test_hdiff_imatch_style.png
