+++ tests/subtests/subtest/baseline/test_hmatch_idiffshape.png
