+++ tests/subtests/subtest/baseline/test_hmissing_idiffshape.png
