+++ tests/subtests/subtest/pytest.ini
@@ -1,0 +1,8 @@
+[pytest]
+markers =
+    image: run test during image comparison only mode.
+    hash: run test during hash comparison only mode.
+filterwarnings =
+    error
+    ignore:distutils Version classes are deprecated
+    ignore:the imp module is deprecated in favour of importlib
