+++ tests/subtests/summaries/test_html_hashes_only.json
@@ -1,5 +1,5 @@
 {
-  "subtests.subtest.test_hmatch_imatch": {
+  "subtests.subtest.test_functions.test_hmatch_imatch": {
     "status": "passed",
     "image_status": null,
     "hash_status": "match",
@@ -8,98 +8,98 @@
     "diff_image": null,
     "rms": null,
     "tolerance": null,
-    "result_image": "subtests.subtest.test_hmatch_imatch/result.png",
+    "result_image": "subtests.subtest.test_functions.test_hmatch_imatch/result.png",
     "baseline_hash": "###_BASELINE_HASH_###",
     "result_hash": "###_RESULT_HASH_###"
   },
-  "subtests.subtest.test_hdiff_imatch": {
+  "subtests.subtest.test_functions.test_hdiff_imatch": {
     "status": "failed",
     "image_status": null,
     "hash_status": "diff",
-    "status_msg": "REGEX:Hash ###_RESULT_HASH_### doesn't match hash ###_BASELINE_HASH_### in library .*\\.json for test subtests\\.subtest\\.test_hdiff_imatch\\.",
+    "status_msg": "REGEX:Hash ###_RESULT_HASH_### doesn't match hash ###_BASELINE_HASH_### in library .*\\.json for test subtests\\.subtest\\.test_functions\\.test_hdiff_imatch\\.",
     "baseline_image": null,
     "diff_image": null,
     "rms": null,
     "tolerance": null,
-    "result_image": "subtests.subtest.test_hdiff_imatch/result.png",
+    "result_image": "subtests.subtest.test_functions.test_hdiff_imatch/result.png",
     "baseline_hash": "###_BASELINE_HASH_###",
     "result_hash": "###_RESULT_HASH_###"
   },
-  "subtests.subtest.test_hmissing_imatch": {
+  "subtests.subtest.test_functions.test_hmissing_imatch": {
     "status": "failed",
     "image_status": null,
     "hash_status": "missing",
-    "status_msg": "REGEX:Hash for test 'subtests\\.subtest\\.test_hmissing_imatch' not found in .*\\.json\\. Generated hash is ###_RESULT_HASH_###\\.",
+    "status_msg": "REGEX:Hash for test 'subtests\\.subtest\\.test_functions\\.test_hmissing_imatch' not found in .*\\.json\\. Generated hash is ###_RESULT_HASH_###\\.",
     "baseline_image": null,
     "diff_image": null,
     "rms": null,
     "tolerance": null,
-    "result_image": "subtests.subtest.test_hmissing_imatch/result.png",
+    "result_image": "subtests.subtest.test_functions.test_hmissing_imatch/result.png",
     "baseline_hash": null,
     "result_hash": "###_RESULT_HASH_###"
   },
-  "subtests.subtest.test_hdiff_imatch_tolerance": {
+  "subtests.subtest.test_special.test_hdiff_imatch_tolerance": {
     "status": "failed",
     "image_status": null,
     "hash_status": "diff",
-    "status_msg": "REGEX:Hash ###_RESULT_HASH_### doesn't match hash ###_BASELINE_HASH_### in library .*\\.json for test subtests\\.subtest\\.test_hdiff_imatch_tolerance\\.",
+    "status_msg": "REGEX:Hash ###_RESULT_HASH_### doesn't match hash ###_BASELINE_HASH_### in library .*\\.json for test subtests\\.subtest\\.test_special\\.test_hdiff_imatch_tolerance\\.",
     "baseline_image": null,
     "diff_image": null,
     "rms": null,
     "tolerance": null,
-    "result_image": "subtests.subtest.test_hdiff_imatch_tolerance/result.png",
+    "result_image": "subtests.subtest.test_special.test_hdiff_imatch_tolerance/result.png",
     "baseline_hash": "###_BASELINE_HASH_###",
     "result_hash": "###_RESULT_HASH_###"
   },
-  "subtests.subtest.test_hdiff_idiff_tolerance": {
+  "subtests.subtest.test_special.test_hdiff_idiff_tolerance": {
     "status": "failed",
     "image_status": null,
     "hash_status": "diff",
-    "status_msg": "REGEX:Hash ###_RESULT_HASH_### doesn't match hash ###_BASELINE_HASH_### in library .*\\.json for test subtests\\.subtest\\.test_hdiff_idiff_tolerance\\.",
+    "status_msg": "REGEX:Hash ###_RESULT_HASH_### doesn't match hash ###_BASELINE_HASH_### in library .*\\.json for test subtests\\.subtest\\.test_special\\.test_hdiff_idiff_tolerance\\.",
     "baseline_image": null,
     "diff_image": null,
     "rms": null,
     "tolerance": null,
-    "result_image": "subtests.subtest.test_hdiff_idiff_tolerance/result.png",
+    "result_image": "subtests.subtest.test_special.test_hdiff_idiff_tolerance/result.png",
     "baseline_hash": "###_BASELINE_HASH_###",
     "result_hash": "###_RESULT_HASH_###"
   },
-  "subtests.subtest.test_hdiff_imatch_savefig": {
+  "subtests.subtest.test_special.test_hdiff_imatch_savefig": {
     "status": "failed",
     "image_status": null,
     "hash_status": "diff",
-    "status_msg": "REGEX:Hash ###_RESULT_HASH_### doesn't match hash ###_BASELINE_HASH_### in library .*\\.json for test subtests\\.subtest\\.test_hdiff_imatch_savefig\\.",
+    "status_msg": "REGEX:Hash ###_RESULT_HASH_### doesn't match hash ###_BASELINE_HASH_### in library .*\\.json for test subtests\\.subtest\\.test_special\\.test_hdiff_imatch_savefig\\.",
     "baseline_image": null,
     "diff_image": null,
     "rms": null,
     "tolerance": null,
-    "result_image": "subtests.subtest.test_hdiff_imatch_savefig/result.png",
+    "result_image": "subtests.subtest.test_special.test_hdiff_imatch_savefig/result.png",
     "baseline_hash": "###_BASELINE_HASH_###",
     "result_hash": "###_RESULT_HASH_###"
   },
-  "subtests.subtest.test_hdiff_imatch_style": {
+  "subtests.subtest.test_special.test_hdiff_imatch_style": {
     "status": "failed",
     "image_status": null,
     "hash_status": "diff",
-    "status_msg": "REGEX:Hash ###_RESULT_HASH_### doesn't match hash ###_BASELINE_HASH_### in library .*\\.json for test subtests\\.subtest\\.test_hdiff_imatch_style\\.",
+    "status_msg": "REGEX:Hash ###_RESULT_HASH_### doesn't match hash ###_BASELINE_HASH_### in library .*\\.json for test subtests\\.subtest\\.test_special\\.test_hdiff_imatch_style\\.",
     "baseline_image": null,
     "diff_image": null,
     "rms": null,
     "tolerance": null,
-    "result_image": "subtests.subtest.test_hdiff_imatch_style/result.png",
+    "result_image": "subtests.subtest.test_special.test_hdiff_imatch_style/result.png",
     "baseline_hash": "###_BASELINE_HASH_###",
     "result_hash": "###_RESULT_HASH_###"
   },
-  "subtests.subtest.test_hdiff_imatch_removetext": {
+  "subtests.subtest.test_special.test_hdiff_imatch_removetext": {
     "status": "failed",
     "image_status": null,
     "hash_status": "diff",
-    "status_msg": "REGEX:Hash ###_RESULT_HASH_### doesn't match hash ###_BASELINE_HASH_### in library .*\\.json for test subtests\\.subtest\\.test_hdiff_imatch_removetext\\.",
+    "status_msg": "REGEX:Hash ###_RESULT_HASH_### doesn't match hash ###_BASELINE_HASH_### in library .*\\.json for test subtests\\.subtest\\.test_special\\.test_hdiff_imatch_removetext\\.",
     "baseline_image": null,
     "diff_image": null,
     "rms": null,
     "tolerance": null,
-    "result_image": "subtests.subtest.test_hdiff_imatch_removetext/result.png",
+    "result_image": "subtests.subtest.test_special.test_hdiff_imatch_removetext/result.png",
     "baseline_hash": "###_BASELINE_HASH_###",
     "result_hash": "###_RESULT_HASH_###"
   }
